Add a systemd option for the leaderlog
The script will figure out where in the epoch we are and if the leaderlog for
the next epoch is available or not. It should be activated by a systemd timer
daily at 09:45 UTC (instructions in INSTALL.md), but it can be run manually at
any time as well.

The script can be configured to do all or any of the following tasks:
 - Send prev and current slots to PoolTool on day 1 of the epoch
 - Mail the leaderlog for the next epoch on day 4 of the epoch
 - Write a slots.csv file on day 4 of the epoch